refactor: integrate RichTextEditor with cover letter and app routing
