+++ package.json
@@ -9,10 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "html2canvas": "^1.4.1",
-    "jspdf": "^2.5.1",
-    "@mantine/core": "^7.6.1",
-    "@mantine/hooks": "^7.6.1",
     "@tabler/icons-react": "^3.0.0",
     "@tiptap/extension-link": "^2.2.4",
     "@tiptap/pm": "^2.2.4",
@@ -20,6 +16,8 @@
     "@tiptap/starter-kit": "^2.2.4",
     "clsx": "^2.1.0",
     "framer-motion": "^11.0.8",
+    "html2canvas": "^1.4.1",
+    "jspdf": "^2.5.1",
     "next": "14.1.2",
     "react": "^18",
     "react-dom": "^18",
@@ -36,7 +34,6 @@
     "eslint": "^8",
     "eslint-config-next": "14.1.2",
     "postcss": "^8.4.35",
-    "postcss-preset-mantine": "^1.13.0",
     "postcss-simple-vars": "^7.0.1",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
