+++ src/app/globals.css
@@ -25,11 +25,11 @@ body {
 }
 
 body {
-  /* color: rgb(var(--foreground-rgb)); */
+  color: rgb(var(--foreground-rgb));
   background-color: black;
-  /* background: linear-gradient(to bottom,
+  background: linear-gradient(to bottom,
       transparent,
-      rgb(var(--background-end-rgb))) rgb(var(--background-start-rgb)); */
+      rgb(var(--background-end-rgb))) rgb(var(--background-start-rgb));
 }
 
 @layer utilities {
