+++ src/app/layout.tsx
@@ -1,10 +1,6 @@
 import "./globals.css";
-import "@mantine/core/styles.css";
-import "@mantine/tiptap/styles.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import { MantineProvider, ColorSchemeScript } from "@mantine/core";
-import { theme } from "./theme";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -20,13 +16,10 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en" className="dark">
-      <head>
-        <ColorSchemeScript defaultColorScheme="dark" />
-      </head>
       <body className={inter.className}>
-        <MantineProvider defaultColorScheme="dark" theme={theme}>
-          <main className="dark text-foreground bg-background">{children}</main>
-        </MantineProvider>
+        <main className="h-full dark text-foreground bg-background">
+          {children}
+        </main>
       </body>
     </html>
   );
