--- src/components/TextEditor.tsx
@@ -1,29 +1,0 @@
-"use client";
-
-import { RichTextEditor } from "@mantine/tiptap";
-import { useEditor } from "@tiptap/react";
-import StarterKit from "@tiptap/starter-kit";
-
-export function TextEditor({ content }: { content?: string }) {
-  const editor = useEditor({
-    extensions: [StarterKit],
-    content,
-  });
-
-  return (
-    <RichTextEditor editor={editor}>
-      <RichTextEditor.Toolbar sticky>
-        <RichTextEditor.ControlsGroup>
-          <RichTextEditor.Bold />
-        </RichTextEditor.ControlsGroup>
-      </RichTextEditor.Toolbar>
-      <RichTextEditor.Content
-        style={{
-          maxHeight: "calc(100vh - 200px)",
-          overflowY: "auto",
-          overflowX: "hidden",
-        }}
-      />
-    </RichTextEditor>
-  );
-}
